+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Composes the DS `Tabs` component from Chakra's `Tabs`, `TabList`, `Tab`, `TabPanels`, and `TabPanel` components.
 - Moved SCSS and CSS variable styles into a custom NYPL theme to extend Chakra's base theme.
 - Updates the `errorText` prop name to `invalidText`.
+- Composes the DS `Select` component from Chakra's `Select` component.
 
 ### Removes